Simplify code by eliminating abstractions.
Gerrit emoji settings don't need to be configurable, because this way everyone
will use the same "style" (basically the same set of emojis) and everyone will
be on the same page, so one team will see the exact same symbols on the other
team's channel. Or if it needs to be changed, it will change for everybody.

Also it's not a command line application anymore, don't need CLI interface.