Add reaction to review request messages
instead of reply in a thread. Moved the now-unnecessary functions to async API
for later possible use.

Implemented add reaction method in async manner, so it can be used in tandem
with the RTM api. For that to work, had to reimplement _get() and _post()
methods in a coroutine, because requests would block here.

The RTM api related calls are prefixed with rtm_ for now.