Rewrite HTTP request lib to aiohttp and completely wipe out requests
I don't want to use two libraries for the exact same purpose (aiohttp and
requests) and I hate requests, so this is a good opportunity to get rid of it.

This way, only one kind of methods should be maintained (async _get and _post)
and when synchrounous calls will suffice, they need to simply run in
the event loop.

The public interfaces are not changed at all, everything seemingly works
the same as before with requests. The async methods has been moved to _ApiBase,
which has a small helper method _run() (which is just a shorthand really).